chore(tests): move dismiss reviews action tests in their own file
