+++ mergify_engine/tests/functional/test_engine.py
@@ -168,78 +168,6 @@ class TestEngineV2Scenario(base.FunctionalTestBase):
         p, _ = self.create_pr()
         debug.report(p.html_url)
 
-    def test_dismiss_reviews(self):
-        return self._test_dismiss_reviews()
-
-    def test_dismiss_reviews_custom_message(self):
-        return self._test_dismiss_reviews(message="Loser")
-
-    def _test_dismiss_reviews(self, message=None):
-        rules = {
-            "pull_request_rules": [
-                {
-                    "name": "dismiss reviews",
-                    "conditions": [f"base={self.master_branch_name}"],
-                    "actions": {
-                        "dismiss_reviews": {
-                            "approved": True,
-                            "changes_requested": ["mergify-test1"],
-                        }
-                    },
-                }
-            ]
-        }
-
-        if message is not None:
-            rules["pull_request_rules"][0]["actions"]["dismiss_reviews"][
-                "message"
-            ] = message
-
-        self.setup_repo(yaml.dump(rules))
-        p, commits = self.create_pr()
-        branch = self.get_full_branch_name("fork/pr%d" % self.pr_counter)
-        self.create_review(p, commits[-1], "APPROVE")
-
-        self.assertEqual(
-            [("APPROVED", "mergify-test1")],
-            [(r.state, r.user.login) for r in p.get_reviews()],
-        )
-
-        open(self.git.tmp + "/unwanted_changes", "wb").close()
-        self.git("add", self.git.tmp + "/unwanted_changes")
-        self.git("commit", "--no-edit", "-m", "unwanted_changes")
-        self.git("push", "--quiet", "fork", branch)
-
-        self.wait_for("pull_request", {"action": "synchronize"})
-        self.wait_for("pull_request_review", {"action": "dismissed"})
-
-        self.assertEqual(
-            [("DISMISSED", "mergify-test1")],
-            [(r.state, r.user.login) for r in p.get_reviews()],
-        )
-
-        commits = list(p.get_commits())
-        self.create_review(p, commits[-1], "REQUEST_CHANGES")
-
-        self.assertEqual(
-            [("DISMISSED", "mergify-test1"), ("CHANGES_REQUESTED", "mergify-test1")],
-            [(r.state, r.user.login) for r in p.get_reviews()],
-        )
-
-        open(self.git.tmp + "/unwanted_changes2", "wb").close()
-        self.git("add", self.git.tmp + "/unwanted_changes2")
-        self.git("commit", "--no-edit", "-m", "unwanted_changes2")
-        self.git("push", "--quiet", "fork", branch)
-
-        self.wait_for("pull_request", {"action": "synchronize"})
-        self.wait_for("pull_request_review", {"action": "dismissed"})
-
-        # There's no way to retrieve the dismiss message :(
-        self.assertEqual(
-            [("DISMISSED", "mergify-test1"), ("DISMISSED", "mergify-test1")],
-            [(r.state, r.user.login) for r in p.get_reviews()],
-        )
-
     def test_backport_no_branch(self):
         rules = {
             "pull_request_rules": [
